requirements.txt
# Core dependencies for UML Evaluation Framework

# Sentence transformers for SBERT embeddings
sentence-transformers>=2.2.0

# OpenAI-compatible API client (works with OpenRouter)
openai>=1.0.0

# Scientific computing
numpy>=1.24.0
scipy>=1.10.0
pandas>=2.0.0

# For structured data handling
pydantic>=2.0.0

# Environment variable management
python-dotenv>=1.0.0

# Progress bars for long operations
tqdm>=4.65.0

# For JSON schema validation
jsonschema>=4.17.0

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0  # For coverage reports
